requirements.txt
playwright == 1.40.0
asyncio
feedgen
TikTokApi
config
google-cloud-storage
